Renamed NodeDataProvider to StateProvider
