+++ packages/recoil-state-provider/src/RecoilStateProvider.test.tsx
@@ -1,7 +1,7 @@
 import { render } from "@testing-library/react";
 import RecoilStateProvider from "./RecoilStateProvider";
 import React from "react";
-import { NodeDataProvider, NodeValidationProvider } from "@graphter/renderer-react";
+import { StateProvider, NodeValidationProvider } from "@graphter/renderer-react";
 import { useRecoilAggregateNodeValidation } from "./validation/useRecoilAggregateNodeValidation";
 import { useRecoilNodeValidation } from "./validation/useRecoilNodeValidation";
 import { NodeValidatorRegistration } from "@graphter/core";
@@ -13,13 +13,13 @@ import { useRecoilTreePaths } from "./data/useRecoilTreePaths";
 import { useRecoilArrayNodeData } from "./data/useRecoilArrayNodeData";
 
 const NodeValidationProviderMock = NodeValidationProvider as jest.Mock<any>
-const NodeDataProviderMock = NodeDataProvider as jest.Mock<any>
+const StateProviderMock = StateProvider as jest.Mock<any>
 
 describe('<RecoilStateProvider />', () => {
   it('should correctly set up Recoil state management', () => {
     const validatorRegistry: Array<NodeValidatorRegistration> = []
     const { container } = render(
-      <RecoilStateProvider instanceId={'some-id'} validatorRegistry={validatorRegistry}>
+      <RecoilStateProvider validatorRegistry={validatorRegistry}>
         <div>Some child</div>
       </RecoilStateProvider>
     )
@@ -31,16 +31,16 @@ describe('<RecoilStateProvider />', () => {
     expect(nodeValidationProviderProps.aggregateNodeValidationHook).toBe(useRecoilAggregateNodeValidation)
     expect(nodeValidationProviderProps.validatorRegistry).toBe(validatorRegistry)
     expect(Object.keys(nodeValidationProviderProps)).toMatchSnapshot()
-    expect(NodeDataProviderMock).toHaveBeenCalled()
-    const nodeDataProviderProps = NodeDataProviderMock.mock.calls[0][0]
-    expect(nodeDataProviderProps.instanceId).toBe('some-id')
-    expect(nodeDataProviderProps.treeDataInitialiserHook).toBe(useRecoilTreeDataInitialiser)
-    expect(nodeDataProviderProps.nodeDataHook).toBe(useRecoilNodeData)
-    expect(nodeDataProviderProps.treeDataHook).toBe(useRecoilTreeData)
-    expect(nodeDataProviderProps.treeDataCallbackHook).toBe(useRecoilTreeDataCallback)
-    expect(nodeDataProviderProps.treePathsHook).toBe(useRecoilTreePaths)
-    expect(nodeDataProviderProps.arrayNodeDataHook).toBe(useRecoilArrayNodeData)
-    expect(Object.keys(nodeDataProviderProps)).toMatchSnapshot()
+    expect(StateProviderMock).toHaveBeenCalled()
+    const stateProviderProps = StateProviderMock.mock.calls[0][0]
+    expect(stateProviderProps.instanceId).toBe('some-id')
+    expect(stateProviderProps.treeDataInitialiserHook).toBe(useRecoilTreeDataInitialiser)
+    expect(stateProviderProps.nodeDataHook).toBe(useRecoilNodeData)
+    expect(stateProviderProps.treeDataHook).toBe(useRecoilTreeData)
+    expect(stateProviderProps.treeDataCallbackHook).toBe(useRecoilTreeDataCallback)
+    expect(stateProviderProps.treePathsHook).toBe(useRecoilTreePaths)
+    expect(stateProviderProps.arrayNodeDataHook).toBe(useRecoilArrayNodeData)
+    expect(Object.keys(stateProviderProps)).toMatchSnapshot()
 
   })
 })
