+++ packages/renderer-react/src/edit-renderer/ValidationSummary.test.tsx
@@ -1,7 +1,7 @@
 import { render } from "@testing-library/react";
 import ValidationSummary from "./ValidationSummary";
 import { NodeValidationProvider } from "../providers/node-validation";
-import NodeDataProvider from "../providers/node-data";
+import StateProvider from "../providers/state";
 import { when } from "jest-when";
 import React from "react";
 
@@ -33,9 +33,10 @@ describe('<ValidationSummary />', () => {
         }
       ])
     const {queryByText} = render(
-      <NodeDataProvider
+      <StateProvider
         treeDataInitialiserHook={jest.fn()}
         nodeDataHook={jest.fn()}
+        multipleNodeDataHook={jest.fn()}
         arrayNodeDataHook={jest.fn()}
         treeDataHook={jest.fn()}
         treeDataCallbackHook={jest.fn()}
@@ -48,7 +49,7 @@ describe('<ValidationSummary />', () => {
         >
           <ValidationSummary config={config} path={[ 'page' ]}/>
         </NodeValidationProvider>
-      </NodeDataProvider>
+      </StateProvider>
     )
     expect(queryByText('page/0: This is an error')).not.toBeNull()
     expect(queryByText('page/0: This is another error')).not.toBeNull()
@@ -61,9 +62,10 @@ describe('<ValidationSummary />', () => {
       type: 'object'
     }
     const { container } = render(
-      <NodeDataProvider
+      <StateProvider
         treeDataInitialiserHook={jest.fn()}
         nodeDataHook={jest.fn()}
+        multipleNodeDataHook={jest.fn()}
         arrayNodeDataHook={jest.fn()}
         treeDataHook={jest.fn()}
         treeDataCallbackHook={jest.fn()}
@@ -76,7 +78,7 @@ describe('<ValidationSummary />', () => {
         >
           <ValidationSummary config={config} path={[ 'page' ]}/>
         </NodeValidationProvider>
-      </NodeDataProvider>
+      </StateProvider>
     )
     expect(container.childElementCount).toBe(0)
   })
