+++ packages/renderer-react/src/index.ts
@@ -2,7 +2,7 @@ import ListRenderer from './list-renderer'
 import EditRenderer from './edit-renderer'
 import NodeEditRenderer from "./edit-renderer/NodeEditRenderer"
 import ValidationSummary from './edit-renderer/ValidationSummary'
-import NodeDataProvider from "./providers/node-data"
+import StateProvider from "./providers/state"
 import ServiceProvider from './providers/service'
 import ConfigProvider from './providers/config'
 
@@ -13,7 +13,7 @@ export {
   ServiceProvider,
   ConfigProvider,
   ValidationSummary,
-  NodeDataProvider,
+  StateProvider,
 }
 
 export * from "./store/nodeRendererStore"
@@ -23,7 +23,7 @@ export * from './util/node'
 export * from './util/path'
 export * from './util/config'
 export * from './providers/node-validation'
-export * from './providers/node-data'
+export * from './providers/state'
 export * from './providers/service'
 export * from './providers/config'
 export * from './setup-node-renderer'
